services/api/requirements.txt
# Streamlit (CRITICAL - must be installed!)
streamlit>=1.28.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Core Dependencies
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
pytrends>=4.9.2
urllib3<2.0  # keep pytrends retry compatibility (method_whitelist)

# Azure SDKs
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-storage-queue>=12.7.0
azure-data-tables>=12.4.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-core>=1.29.5

# AI & ML
openai>=1.12.0

# Utilities
beautifulsoup4>=4.12.0
colorama>=0.4.6
google-ads==28.4.1
PyJWT[crypto]>=2.8.0
